handle common errors

+++ src/routes/deletePokemon.js
@@ -2,15 +2,25 @@ const { Pokemon } = require('../db/sequelize')
   
 module.exports = (app) => {
   app.delete('/api/pokemons/:id', (req, res) => {
-    Pokemon.findByPk(req.params.id).then(pokemon => {
-      const pokemonDeleted = pokemon;
-      Pokemon.destroy({
-        where: { id: pokemon.id }
+    Pokemon.findByPk(req.params.id)
+      .then(pokemon => {
+        if(pokemon === null){
+          const message = 'Pokemon do not exist. Try with an another ID.'
+          return res.status(404).json({message})
+        }
+        
+        const pokemonDeleted = pokemon;
+        return Pokemon.destroy({
+          where: { id: pokemon.id }
+        })
+          .then(_ => {
+            const message = `Pokemon with id n°${pokemonDeleted.id} has been deleted.`
+            res.json({ message, data: pokemonDeleted })
+          })
       })
-      .then(_ => {
-        const message = `Pokemon with id n°${pokemonDeleted.id} has been deleted.`
-        res.json({message, data: pokemonDeleted })
+      .catch(error => {
+        const message = "The Pokemon cannot be deleted. Try later."
+        res.status(500).json({ message, data: error })
       })
-    })
   })
 }
